Coding File of Time series forecasting
Autoregressive Integrated Moving Average Model. An ARIMA model is a class of statistical models for analyzing and forecasting time series data. It explicitly caters to a suite of standard structures in time-series data, and as such provides a simple yet powerful method for making skillful time-series forecasts. plotting all graphs of stationary and non-stationary plots gives you an idea about the next approach to solve the TS problems.